adds font awesome
