+++ react-app/public/index.html
@@ -1,10 +1,14 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <title>Simple React App</title>
-  </head>
-  <body>
-    <div id="root"></div>
-  </body>
+
+<head>
+  <meta charset="utf-8" />
+  <title>Spooky Home Rentals - ScareBnB</title>
+  <script src="https://kit.fontawesome.com/e145510635.js" crossorigin="anonymous"></script>
+</head>
+
+<body>
+  <div id="root"></div>
+</body>
+
 </html>
