--- react-app/src/components/NavBar.js
@@ -1,44 +1,0 @@
-import React from 'react';
-import { NavLink } from 'react-router-dom';
-import LogoutButton from './auth/LogoutButton';
-import LoginFormModal from './auth/LoginFormModal'
-import SignUpFormModal from './auth/SignUpFormModal'
-
-const NavBar = ({ authenticated, setAuthenticated }) => {
-  return (
-    <nav>
-      <ul>
-        <li>
-          <NavLink to="/" exact={true} activeClassName="active">
-            Home
-          </NavLink>
-        </li>
-        <li>
-          <LoginFormModal
-            authenticated={authenticated}
-            setAuthenticated={setAuthenticated}
-          />
-        </li>
-        <li>
-          <SignUpFormModal
-            authenticated={authenticated}
-            setAuthenticated={setAuthenticated}
-          />
-          {/* <NavLink to="/sign-up" exact={true} activeClassName="active">
-            Sign Up
-          </NavLink> */}
-        </li>
-        <li>
-          <NavLink to="/users" exact={true} activeClassName="active">
-            Users
-          </NavLink>
-        </li>
-        <li>
-          <LogoutButton setAuthenticated={setAuthenticated} />
-        </li>
-      </ul>
-    </nav>
-  );
-}
-
-export default NavBar;
