+++ react-app/src/components/SplashPage/HomePage.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function HomePage() {
+  return (
+    <div style={{ backgroundColor: "orange", width: "100vw", height: "100vw", margin: "0px", padding: "0px" }}>
+      Content
+    </div>
+  )
+}
