+++ react-app/src/components/SplashPage/NavBar/NavBar.css
@@ -1,0 +1,4 @@
+nav{
+  display: flex;
+  justify-content: space-around;
+}
