+++ react-app/src/components/SplashPage/NavBar/UserMenu/UserMenu.css
@@ -1,0 +1,3 @@
+.usermenu__dropdown_button {
+
+}
