+++ react-app/src/components/SplashPage/User.js
