+++ react-app/src/components/SplashPage/auth/LoginFormModal/LoginForm.js
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { Redirect } from "react-router-dom";
-import { login } from "../../../services/auth";
+import { login } from "../../../../services/auth";
 
 const LoginForm = ({ authenticated, setAuthenticated }) => {
   const [errors, setErrors] = useState([]);
