+++ react-app/src/components/SplashPage/auth/LoginFormModal/index.js
