+++ react-app/src/components/SplashPage/auth/ProtectedRoute.js
