+++ react-app/src/components/SplashPage/auth/SignUpFormModal/SignUpForm.js
@@ -1,9 +1,8 @@
 import React, { useState } from "react";
 import { Redirect } from 'react-router-dom';
-import { signUp } from '../../../services/auth';
+import { signUp } from '../../../../services/auth';
 
 const SignUpForm = ({ authenticated, setAuthenticated }) => {
-  const [username, setUsername] = useState("");
   const [firstName, setFirstName] = useState("");
   const [lastName, setLastName] = useState("");
   const [bio, setBio] = useState("");
