+++ react-app/src/components/SplashPage/auth/SignUpFormModal/index.js
